move reputation to the main module
