+++ ray_on_golem/reputation/__init__.py
@@ -1,0 +1,3 @@
+from ray_on_golem.reputation.service import ReputationService
+
+__all__ = ("ReputationService",)
