+++ ray_on_golem/reputation/cli.py
@@ -7,9 +7,9 @@ from prettytable import PrettyTable
 from tortoise.exceptions import DoesNotExist
 
 from ray_on_golem.cli import with_datadir
-from ray_on_golem.server.services.reputation import models as m
-from ray_on_golem.server.services.reputation.service import ReputationService
-from ray_on_golem.server.services.reputation.updater import ReputationUpdater
+from ray_on_golem.reputation import models as m
+from ray_on_golem.reputation.service import ReputationService
+from ray_on_golem.reputation.updater import ReputationUpdater
 
 
 @click.group(
@@ -157,20 +157,3 @@ def update(datadir, network):
         )
 
     asyncio.run(_update())
-
-
-@reputation_cli.command()
-@with_network
-@with_datadir
-def test(datadir, network):
-    async def _test():
-        async with ReputationService(datadir):
-            cnt_added, cnt_updated, cnt_ignored, cnt_total = await ReputationUpdater(
-                network
-            ).update(partial(click.progressbar, label="Updating scores"))
-
-        print(
-            f"Reputation DB updated. Total scores={cnt_total} (added={cnt_added}, updated={cnt_updated}, ignored={cnt_ignored})."
-        )
-
-    asyncio.run(_testx())
