+++ ray_on_golem/reputation/plugins/__init__.py
@@ -1,0 +1,4 @@
+from ray_on_golem.reputation.plugins.blacklist import ProviderBlacklistPlugin
+from ray_on_golem.reputation.plugins.score import ReputationScorer
+
+__all__ = ("ProviderBlacklistPlugin", "ReputationScorer")
