+++ ray_on_golem/server/services/__init__.py
@@ -1,12 +1,10 @@
 from ray_on_golem.server.services.golem import GolemService, get_manifest
 from ray_on_golem.server.services.ray import RayService
-from ray_on_golem.server.services.reputation import ReputationService
 from ray_on_golem.server.services.yagna import YagnaService
 
 __all__ = (
     "GolemService",
     "RayService",
-    "ReputationService",
     "YagnaService",
     "get_manifest",
 )
