--- ray_on_golem/server/services/golem/provider_data.py
@@ -1,11 +1,0 @@
-# TODO: remove this file
-
-from typing import Dict, List
-
-# TODO: add PROVIDERS_SCORED VVV
-
-PROVIDERS_SCORED: Dict[str, List[str]] = {
-    "polygon": [
-        "0x7d529a5b0b58400d09154b47d94ee022ba7658cc",
-    ]
-}
