--- ray_on_golem/server/services/reputation/__init__.py
@@ -1,3 +1,0 @@
-from ray_on_golem.server.services.reputation.service import ReputationService
-
-__all__ = ("ReputationService",)
