--- ray_on_golem/server/services/reputation/plugins/__init__.py
@@ -1,4 +1,0 @@
-from ray_on_golem.server.services.reputation.plugins.blacklist import ProviderBlacklistPlugin
-from ray_on_golem.server.services.reputation.plugins.score import ReputationScorer
-
-__all__ = ("ProviderBlacklistPlugin", "ReputationScorer")
